# server/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Project Manga",
        "description": "",
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "tags": [
        {
            "name": "manga",
            "description": "Project Manga Bacth 9",
            "externalDocs": {
                "description": "Find out more",
                "url": "http://swagger.io"
            }
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/manga/": {
            "post": {
                "tags": [
                    "manga"
                ],
                "summary": "Add a new manga",
                "description": "",
                "operationId": "addManga",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Manga object create to db",
                        "required": true
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "tags": [
                    "manga"
                ],
                "summary": "Finds Manga by status",
                "description": "Find all data manga",
                "operationId": "find all manga",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/manga/{id}": {
            "put": {
                "tags": [
                    "manga"
                ],
                "summary": "Update an existing manga",
                "description": "",
                "operationId": "updateManga",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update Manga ",
                        "required": true,
                        "type": "string"
                      },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Manga object create to db",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Region not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "get": {
                "tags": [
                    "manga"
                ],
                "summary": "find an existing manga",
                "description": "",
                "operationId": "updateManga",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Find Manga ",
                        "required": true,
                        "type": "integer"
                      }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Manga not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "manga"
                ],
                "summary": "delete an existing manga",
                "description": "",
                "operationId": "updateManga",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete Manga ",
                        "required": true,
                        "type": "integer"
                      }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Region not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        } 
    }
}